Move all OMEMO related parts to own dir
This will allow to easier split up in to own package. See also
omemo/gajim#22
